get client's ip and move to the port number of model selected by client
