+++ DOCKERFILE/tflite/src/templates/index.html
@@ -28,7 +28,7 @@
     <div class="option">
         <select id="video-transform">
             <option value="tfLite" selected>Tensorflow-lite</option>
-            <option value="other">other</option>
+            <option value="other">mobile-net</option>
         </select>
     </div>
 
@@ -41,8 +41,6 @@
         <audio id="audio" autoplay="true"></audio>
         <video id="video" autoplay="true" playsinline="true"></video>
     </div>
-
-
 <script src="client.js"></script>
 </body>
 </html>
