+++ DOCKERFILE/web/Dockerfile
@@ -9,6 +9,7 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     x11-apps \
     wget \
     unzip \
+    iproute2 \
     && apt-get clean \
     && rm -rf /var/lib/apt/lists/*
 
@@ -36,20 +37,6 @@ COPY requirements.txt .
 
 RUN pip install --break-system-packages --no-cache-dir -r requirements.txt
 
-
-# Download and unzip the TFLite object detection model
-RUN mkdir -p src/TFLite_model/object_detection/
-
-RUN wget https://www.kaggle.com/api/v1/models/tensorflow/efficientdet/tfLite/lite0-detection-default/1/download -O model.tar.gz \
-    && tar -xvzf model.tar.gz -C src/TFLite_model/object_detection \
-    && rm model.tar.gz \
-    && mv src/TFLite_model/object_detection/1.tflite src/TFLite_model/object_detection/efficientdet.tflite
-
-RUN wget https://www.kaggle.com/api/v1/models/tensorflow/ssd-mobilenet-v1/tfLite/default/1/download -O model.tar.gz \
-    && tar -xvzf model.tar.gz -C src/TFLite_model/object_detection \
-    && rm model.tar.gz \
-    && mv src/TFLite_model/object_detection/1.tflite src/TFLite_model/object_detection/ssd-mobilenet-v1.tflite
-    
 COPY src /app/src
 
 CMD ["python3", "src/app.py"]
