+++ DOCKERFILE/web/requirements.txt
@@ -1,6 +1,4 @@
-opencv-python==4.9.0.80
-tflite-runtime==2.14.0
 aiortc==1.9.0
 aiohttp==3.8.4
-av==10.0.0
 docker
+
