+++ DOCKERFILE/web/src/static/client.js
@@ -1,78 +1,30 @@
-var pc = null;
-var serverUrl = '/webrtc';
-const config = { sdpSemantics: 'unified-plan', iceServers: [{ urls: ['stun:stun.l.google.com:19302'] }] };
-
-async function fetchFromServer(type, data) {
-    const response = await fetch(serverUrl, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(data)
-    });
-    return response.json();
-}
-
-async function negotiate() {
-    // Send a request to create an offer
-    const offerResponse = await fetchFromServer('request',{ 
-            type: 'request', 
-            video_transform: document.getElementById('video-transform').value
-        });
-    
-    if (offerResponse.error) {
-        console.error('Error requesting offer:', offerResponse.error);
-        return;
-    }
-
-    const offer = new RTCSessionDescription({
-        sdp: offerResponse.sdp,
-        type: offerResponse.type
-    });
-
-    await pc.setRemoteDescription(offer);
-
-    // Create an answer
-    const answer = await pc.createAnswer();
-    await pc.setLocalDescription(answer);
-
-    // Send the answer to the server
-    const answerResponse = await fetchFromServer('answer', {
-        type: pc.localDescription.type,
-        sdp: pc.localDescription.sdp,
-        id: offerResponse.id
-    });
-
-    if (answerResponse.error) {
-        console.error('Error sending answer:', answerResponse.error);
+async function getCurrentIP(model) {
+    try {
+        const response = await fetch('/get-ip',{
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({ model }),
+        }); 
+        const data = await response.json();
+        return data.ip; // Return the IP from the response
+    } catch (error) {
+        console.error('Error fetching IP:', error);
+        return null;
     }
 }
 
-function start() {
-    pc = new RTCPeerConnection(config);
 
-    // Handle incoming tracks
-    pc.addEventListener('track', (evt) => {
-        if (evt.track.kind == 'video') {
-            document.getElementById('video').srcObject = evt.streams[0];
-        } else {
-            document.getElementById('audio').srcObject = evt.streams[0];
-        }
-    });
-
-    document.getElementById('start').style.display = 'none';
-    negotiate();
-    document.getElementById('stop').style.display = 'inline-block';
+async function move_tflite() {
+    const ip = await getCurrentIP("tflite");
+    if (ip) {
+        window.location.href = `http://${ip}:80`; // Use the fetched IP
+    } else {
+        console.error('Unable to retrieve IP address');
+    }
 }
 
-function stop() {
-    document.getElementById('stop').style.display = 'none';
 
-    // Close peer connection
-    setTimeout(() => {
-        if (pc) {
-            pc.close();
-            pc = null;
-        }
-    }, 500);
-}
+
+
