+++ DOCKERFILE/web/src/templates/index.html
@@ -25,24 +25,8 @@
 <body>
     <h1>WebRTC- Raspberry pi</h1>
     <h2>Options</h2>
-    <div class="option">
-        <select id="video-transform">
-            <option value="tfLite" selected>Tensorflow-lite</option>
-            <option value="other">other</option>
-        </select>
-    </div>
-
-    <button id="start" onclick="start()">Start</button>
-    <button id="stop" style="display: none" onclick="stop()">Stop</button>
-
-    <div id="media">
-        <h2>Media</h2>
-
-        <audio id="audio" autoplay="true"></audio>
-        <video id="video" autoplay="true" playsinline="true"></video>
-    </div>
-
-
+    <button onclick="move_tflite()">tflite, mobile-net</button>
+    
 <script src="client.js"></script>
 </body>
 </html>
