--- test_fastAPI/DOCKERFILE/README.md
@@ -1,5 +1,0 @@
-# How to manage docker and run main.py
-
-1. set your point to DOCKERFILE folder to use ["cd DOCKERFILE"]
-2. "docker-compose up --build" to build image and run main.py
-
