--- test_fastAPI/DOCKERFILE/docker-compose.yml
@@ -1,32 +1,0 @@
-version: '3.8'
-
-services:
-  webrtc-pi:
-    build:
-      context: web/.
-      dockerfile: Dockerfile 
-    volumes:
-      - /run/udev:/run/udev:ro
-      - /var/run/docker.sock:/var/run/docker.sock # mount for using docker demon
-    networks:
-      - your-network
-    ports:
-      - "8080:8080"
-    privileged: true
-
-  tflite:
-    build:
-      context: tflite/.
-      dockerfile: Dockerfile 
-    volumes:
-      - /run/udev:/run/udev:ro
-    privileged: true
-    container_name: tflite
-    ports:
-      - "8000:8000"
-    networks:
-      - your-network
-
-networks:
-  your-network:
-    driver: bridge
