--- test_fastAPI/README.md
@@ -1,4 +1,0 @@
-# Raspberry-Pi-5-Computer-Vision
-![Untitled](https://github.com/NerdConnection/Raspberry-Pi-5-Computer-Vision/assets/100738404/1d7b75f8-ce6c-47d9-b7e5-7c9a76811c98)
-
-Raspberry Pi 5를 이용한 Computer-Vision-Suite 개발
