--- test_fastAPI/실증적 sw프로젝트/자료/README.md
@@ -1,26 +1,0 @@
-![Untitled](https://github.com/NerdConnection/Raspberry-Pi-5-Computer-Vision/assets/100738404/1d7b75f8-ce6c-47d9-b7e5-7c9a76811c98)
-
-### 프로젝트 순서
-
-1. 문제 정의 먼저 해야함
-    
-    (FairDay 발표 때 첫 마디로 할 말)
-    
-    → 이 기술과 프로젝트가 왜 필요한가
-    
-2. 문제를 해결하기 위한 기술 list up 
-    
-    ⇒ 해당 기술에 대한 reference 조사 
-    
-    → 문제점 찾고, 해결or성능 향상
-    
-3. ~ A, B, C 기술에 이런 장단점이 있어서 
-    
-    우리는 이렇게 개선하고 결합해서 
-    
-    구현or해결 했다.
-
-- 추가적으로 더 복잡한 문제 정의 필요
-    - RasPi의 접근성과 성능을 활용한 CV Benchmark로써의 활용
-    - 기존의 있는 컴퓨터 비전의 문제점을 보완?
-    - 여러 변수들의 고려를 위한 하드웨어 쪽으로의 개발
