--- test_fastAPI/실증적 sw프로젝트/자료/Raspberry Pi에서 빈번하게 일어나는 issue.md
@@ -1,50 +1,0 @@
-**Raspberry pi 에서는 한정된 리소스를 제공하기 때문에 ARM 아키텍처에 특화된 컴퓨터 비전 모델을 사용할 필요가 있다**
-
-
-1. EfficientDet:
-
-높은 성능을 유지하면서도 상대적으로 적은 컴퓨팅 리소스를 사용합니다.
-객체 감지 작업에서 경쟁력 있는 성능을 제공하며, 컴퓨팅 리소스를 효율적으로 활용합니다.
-
-2. MobileNet:
-
-경량화된 컨볼루션 신경망으로, 모바일 및 임베디드 장치에서 사용하기에 적합합니다.
-이미지 분류, 객체 감지 및 세분화와 같은 다양한 컴퓨터 비전 작업에 사용될 수 있습니다.
-모델의 경량화로 인해 상대적으로 적은 컴퓨팅 리소스를 사용하며, 이로써 제한된 환경에서도 효율적으로 실행될 수 있습니다.
-
-3. YOLO (You Only Look Once)
-
-open cv 최소 3.4.2 버전 필요 cpu에서만 작동 비디오 실시간 처리 속도가 느림
-darknet yolo 를 위해 특별히 제작 되었다 gpu cpu 함께 사용 가능 리눅스에서만 호환
-
-4. Tiny YOLO: 
-
-**raspberry pi 와 같이 낮은 성능의 개발보드에 적합  매우좋음
-단점 낮은 해상도 때문에 저조한 인식률 raspberry pi 5에서는 괜찮은 성능을 보여줄수 있음**
-YOLO의 경량 버전으로, 모델의 크기와 계산량을 줄여 리소스 제한된 환경에서 효율적으로 실행될 수 있습니다.
-객체 감지를 위한 경량화된 솔루션으로 인기가 있으며, YOLO보다 더 작은 장치에서도 빠르게 실행될 수 있습니다.
-
-5. SSD (Single Shot Multibox Detector): 
-
-다양한 크기의 객체를 식별하는 데 사용되는 빠르고 효율적인 객체 감지 모델입니다.
-경량화된 버전의 SSD도 사용할 수 있으며, 라즈베리 파이와 같은 제한된 환경에서도 잘 작동합니다.
-객체 감지 작업에서 빠른 실행 속도와 높은 정확도를 제공합니다.
-
-
-**Raspberry pi 에 도커 환경을 설정할때 일어나는 이슈**
-
-1. 하드웨어 리소스 부족: 라즈베리 파이는 일반 컴퓨터 보다 하드웨어 성능이 제한적이므로 도커 컨테이너를 실행할때 리소스가 
-부족할 수 있다. 메모리와 cpu 리소스 관리에 유의해야 한다
-
-2. 아키텍처 호환성: 라즈베리 파이의 ARM 아키텍처는 x86아키텍처와 다르기 때문에 도커 이미지가 호환 되지 않을 수 있다 
-이 경우 **ARM 아키텍처용 이미지를 사용 해야 한다**
-
-3. 네트워크 문제: 네트워크 구성이 잘못되면 도커 컨테이너가 외부와 통신하는 데 문제가 발생할 수 있다
-
-4. 저장소 용량: 도커 이미지와 컨테이너가 저장소를 적절히 활용할 수 있도록 관리.
-
-5. 커널 호환성: 호스트 시스템의 커널 버전이 도커와 호환되지 않을 수 있다. **커널 버전 확인 필수**
-
-6. 시스템 리소스 충돌: 라즈베리 파이에서 다른 서비스나 프로세스가 실행 중인 경우에는 도커가 충돌 할 수 있다.
-필요한 경우 다른 서비스와 충돌하지 않도록 포트나 리소스를 조정해야 한다.
-
