--- test_fastAPI/실증적 sw프로젝트/자료/중간발표초안.md
@@ -1,68 +1,0 @@
-# 1. 현재 진행중인 프로젝트의 정의,목표
-### 임베디드 시스템, 대표적으로 라즈베리 파이에서 다양한 컴퓨터 비전 모델을 테스트하려는 사용자를 대상으로 한다. 이러한 사용자들에게 필요한 환경을 손쉽게 구축할 수 있도록 도와주는 저장소를 만드는 것이 프로젝트의 목표이다.
-</br> </br>
-
-# 2. 이 프로젝트를 왜 해야하는가?
->> - ### '임베디드 시스템에서 Computer Vision 테스트 환경 제공'의 필요성은 크게 두 가지로 나눌 수 있다.
->>  - #### 임베디드 시스템에서의 Computer Vision은 특정 상황이나 환경에서 필수불가결 
->>  -  #### 임베디드 시스템에서의 Computer Vision 수요의 증가  
-
-
-</br>
-
-## (1) 임베디드 시스템에서의 Computer Vision은 특정 상황이나 환경에서 필수불가결 
-- ### 인터넷이 제공되지않거나, 서버를 설치하기 어려운공간에서의 수요
-      1. 재난 현장이나 원격 지역에서 응급 구조를 위해 사용되는 의료 장비에 CV 기술을 적용
-      2. 인터넷이 연결되지 않은 고립된 장소나 보안이 중요한 환경에서 CV를 사용한 보안 시스템을 구축
-
-- ### 엣지 컴퓨팅을 활용한 전처리된 CV 데이터의 서버 전송
-        엣지 컴퓨팅 임베디드 시스템을 활용하여 CV 데이터를 전처리하고, 필요한 정보만을 서버로 안전하게 전송. 이를 통해 서버의 데이터 전처리 부담을 줄이고 실시간 응답성과 데이터 보안을 향상
-    
-        1. 데이터 전송 부담 완화: 대규모의 CV 데이터를 서버로 전송하는 것은 대역폭과 데이터 처리량에 부담을 줄 수 있다. 엣지 컴퓨팅 임베디드 시스템에서 전처리를 수행하고 필요한 정보만을 서버로 전송하면 네트워크 부하를 줄일 수 있습니다. (예시 : 스마트 시티 교통 모니터링 시스템)
-
-        2. 실시간 응답: 일부 애플리케이션은 실시간 응답이 필요합니다. 엣지 임베디드 시스템에서 데이터를 전처리하고 전송하면 서버까지의 지연 시간을 줄일 수 있어 실시간 응답성을 향상시킬 수 있습니다.
-
-        3. 보안 및 개인 정보 보호: 민감한 데이터를 전송할 때 엣지 컴퓨팅을 사용하면 데이터의 보안과 개인 정보 보호를 강화할 수 있습니다. 먼저 데이터를 로컬에서 처리하고 필요한 정보만을 안전한 방식으로 서버에 전송할 수 있습니다. (예시 : 의료 분야에서는 엣지 컴퓨팅을 사용하여 환자의 의료 정보를 로컬에서 처리)
-
->>스마트 시티 교통 모니터링 시스템 예시 (TIMOS)
-http://bitsensing.co.kr/news/mobility/230207/79
-</br>
-
-## (2) 임베디드 시스템의 수요의 증가 
-
-- ### CV을 사용한 임베디드 시스템의 수요 증가
--  
-      1. 임베디드 시장 성장률 시각자료
-      2. CV을 사용한 임베디드 시스템의 자세한 예시 열거
-            -  ADAS(Advanced Driver Assistance Systems)
-            -  스마트 시티 및 교통 시스템 : 신호등 감지, 교차로 효율화, 주차 공간 모니터링
-            -  스마트 홈 및 IoT : 보안 및 에너지 관리 시스템
-
-![image](https://github.com/NerdConnection/Raspberry-Pi-5-Computer-Vision/assets/100738404/b8060797-b979-45a0-a131-012fe0daf0b7)
-
-
-
-</br></br>
-
-# (3) 라즈베리파이 5 
-- ### 2023.09.28 새롭게 출시된 라즈베리파이 5
-      - 대중적으로 많이 보급되어있으며, 최근 출시된 RPi 5 모델은 이전 모델에 비해 GPU가 향상되어 딥러닝 모델을 구동하기에 적합하다.
-      - 다른 모델(Ndivia Jetson)과 비교했을때 비슷한 성능이지만 저렴하다.
-      - 향상된 cpu와 PCIe의 탑재 (cpu성능은 1.5배, gpu성능은 2.5배)
-
-
-
-
-
-SeeedStudio의 벤치마크 : SeeedStudio는 모바일용 딥 러닝 추론 프레임워크인 ncnn을 사용하고 있는 것이 특징으로, Vulkan API를 통해 GPU 가속을 제공하고 있다는 것으로, Raspberry Pi 5와 Raspberry Pi 4에서 검증한 결과를 그래프로 정리한 것이 이하입니다. 막대그래프가 짧을수록 성능이 뛰어나다는 것을 보여 주며, 어느 쪽이든 Raspberry Pi 5 점수가 Raspberry Pi 4 점수를 초과한다는 것을 알 수 있습니다.
-https://doooob.tistory.com/1483 </br>
-<원문>:https://www.raspberrypi.com/news/benchmarking-raspberry-pi-5/
-![image](https://github.com/NerdConnection/Raspberry-Pi-5-Computer-Vision/assets/100738404/756fabde-e829-4db9-b129-23ab5d8ab6eb)
-
-
-
-</br></br>
-
-
-# 현재 진행상황
-### Raspi OS에 docker를 설치하여 원활한 vision model 실행을 위한 카메라 입력 라이브러리, vision model 프레임워크인 tensorflow lite와 모델 가속화 ONNX runtime 환경을 구축하는 개발을 진행
