--- test_fastAPI/실증적 sw프로젝트/자료/카메라와 AI Framwork (기술, 환경).md
@@ -1,8 +1,0 @@
-https://076923.github.io/posts/Python-opencv-1/ open cv 강좌
-https://076923.github.io/posts/Python-opencv-1/ open cv 강좌 opencv 관련 github 오픈 소스 아키텍처. https://github.com/interactionresearchstudio/NaturewatchCameraServer.git OpenCV를 사용하여 카메라 스트리밍을 웹 브라우저로 전송하는 웹 서버입니다
-
-https://github.com/yushuhuang/webcam.git 웹캠에서 영상을 캡처하여 웹 페이지에 실시간으로 스트리밍하는 간단한 웹 애플리케이션입니다
-
-https://github.com/Mjrovai/OpenCV-Object-Face-Tracking.git OpenCV를 사용하여 물체를 추적하는 프로젝트입니다 다양한 추적 알고리즘을 포함하고 있습니다
-
-https://github.com/Mjrovai/OpenCV-Face-Recognition.git OpenCV를 사용하여 얼굴을 인식하는 간단한 프로젝트입니다. 사전 훈련된 얼굴 감지기를 사용합니다
