--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-03-18.md
@@ -1,22 +1,0 @@
-# To do
-## 라즈베리파이 5에서 OpenCV를 활용하여 CV 모델을 적용할 때에 생기는 문제들을 개선
-
-<br>
-
-1. OpenCV 에서 제공하는 모델들이 어떤 것이 있는지 (각 모델의 기능과, 입력 출력 모양)
-2. Raspberry Pi에서 Docker 환경을 실행 할 때에 어떤 문제가 있는지 (Docker 활용 방법)
-3. Raspberry Pi에서 빈번하게 일어나는 issue, 제공하는 사양이 얼마나 되는지 (통신 속도, 실행 할 수 있는 모델 크기 한계)
-4. Raspberry Pi에서 정보를 Web으로 넘겨 줄 때 issue (통신 속도, 통신 방법)
-
-### 3/20 13:30에 회의 후 문제를 종합해서 멘토링 신청
-
-#### 조현진
-1. OpenCV 에서 제공하는 모델들이 어떤 것이 있는지 (각 모델의 기능과, 입력 출력 모양)
-
-#### 허남수, 서윤오
-2. Raspberry Pi에서 Docker 환경을 실행 할 때에 어떤 문제가 있는지 (Docker 활용 방법)
-4. Raspberry Pi에서 정보를 Web으로 넘겨 줄 때 issue (통신 속도, 통신 방법)
-
-#### 반석
-3. Raspberry Pi에서 빈번하게 일어나는 issue, 제공하는 사양이 얼마나 되는지 (통신 속도, 실행 할 수 있는 모델 크기 한계)
-
