--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-03-25.md
@@ -1,77 +1,0 @@
-# 멘토 질문 목록
-
-
-1. 프로젝트 목표
-    - 많은 모델을 사용자가 편리하게 사용할 수 있게 환경을 제공
-    - 몇개의 Basic 모델들을 최적화하여, 간단한 동작으로 빠른 연산을 제공
-    - 본인의 모델을 쉽게 적용시켜 비교, 테스트 할 수 있는 프레임워크를 제공
-2. 작성한 서비스 제공 로드맵 검토
-3. 프레임 워크 선정
-    - TensorFlow Lite, NCNN, ONNX Runtime 같은 프레임워크 중 선정 방법, 기준
-4. AI 가속화 PCIe 인터페이스 사용 여부
-    - 사용 한다면, 어떤 기준을 삼고 골라야 하는지
-  
-<hr/>
-
-**AI 가속화를 위한 PCIe 인터페이스**
-
-**Coral PCIe Accelerator: https://coral.ai/products/pcie-accelerator/**
-
-**Hailo-8 M.2 PCIe Accelerator: https://hailo.ai/products/hailo-8-m2-module/**
-
-**Sima.ai MLSoC Dual M.2 Eval Development Kit: https://sima.ai/mlsoc-boards/**
-
----
-
-**Tensorflow Lite (기존 텐서플로 모델→ 텐서플로 라이트로 변환 필요)**
-
-**NCNN (사용법 복잡)**
-
-**ONNX Runtime (텐서플로 → ONNX 변환 필요)**
-
-**성능 ⇒ NCNN**
-
-**사용 편의성 ⇒ TF Lite**
-
-**성능, 다양한 모델 ⇒ ONNX Runtime**
-
----
-
-**최대한 많은 모델을 담는 것?**
-
-**몇개의 모델에 대해 빠른 연산과 편리한 사용을 제공하는 것?**
-
-**본인의 모델을 쉽게 적용시킬 수 있게 프레임워크를 제공하는 것?**
-
----
-# 피드백 
-
- **최대한 많은 모델을 담는 것? --> 간단한 cv 모델들을 담는 것**   
- 
- **몇개의 모델에 대해 빠른 연산과 편리한 사용을 제공하는 것? --> 우선 하나의 모델을 최적화**   
-   
- **본인의 모델을 쉽게 적용시킬 수 있게 프레임워크를 제공하는 것? --> yes, 몇가지 모델 선정 후 고민**     
- - 프레임 워크 onnix(멘토님 추천)
-
- **AI 가속화를 위한 PCIe 인터페이스 사용이 필수인지? --> 사용하지 않아도 됨**
-    
----   
-    
- **진행 목표**    
-   - 누구나 라즈베리 파이가 있다면 이 시스템을 사용해서 돌릴 수 있는 시스템으로 먼저 개발
-     웹서비스로 할 것이라면 라즈베리파이 내에서 돌아가는 작은 서버를 활용할 것.
-     기존의 방식은 임베디드에서 멀어지고 IOT에 더 가까움.
-        
-  - 진행방향 : 라즈베리 키트에 있는 카메라를 호환하는 것을 첫 번째 목표로 설정하고 후에 성능이 더 뛰어난 카메                라 쪽으로 세팅   
-              ( ex. arducam )      
-  - 사용자들이 필요한 부분이 있다면 사용할 수 있는 github프로젝트를 작성
-      1. 사용하는 방법 안내 (가이드)
-      2. 모든 개발과정을 issue로 관리
-      3. 분석 결과등이 담긴 작은 규모의 성공한 레파지토리들을 미리 참고
-
----   
-       
-  **구현 후 방향**    
- - 여러가지에 대응하는 외부 web에서 사용한다거나 다른 모델들도 호환이 되도록 하는 CV suite개발
-   
-
