--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-04-01.md
@@ -1,15 +1,0 @@
-## 해야 할 일 ##
-
-## 조현진
-- 개발 배경, 문제 정의 및 reference 정리
-- baseline으로 들어갈 vision model 구현
-
-## 서버 팀 (서윤오, 허남수)
-- Raspi에 Docker 환경 구현 해 보고, 공부해서 팀원 알려주기
-
-## 임베디드 팀(반석, 나영준)
-- 재료비로 물건 구매 청구서 메일 보내기
-- Raspi에서 사용하는 Framework (ONXX Runtime) 익히기
-
-# 공통사항
-- github repo 관리를 위한 git 공부
