--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-04-08.md
@@ -1,26 +1,0 @@
-# 지난주 회고
-   ### 허남수, 서윤오
-     - 도커 설명을 위한 자료 작성
-     - 도커환경에서 YOLOV5 모델 실행 후 결과 확인
-     - 가중치 파일이 추가된 YOLOV5 이미지 모델을 빌드하는 기초적인 DOCKERFILE 작성
-   ### 조현진
-     - 프로젝트 문제 정의 재작성
-     - Vision model, OpenCV 자료조사
-     
-   ### 반석, 나영준
-     - onnx를 통해 open cv, scikit learn로 작성된 코드 onnx로 변환 후 onnx runtime을 통해 실행
-
- 
- # 이번주 할 일
-   ### 허남수, 서윤오
-     - 라즈베리 파이5에서 도커환경 구축 및 cv모델을 위한 기본이미지 생성, 이미지가 베포 효용성 확인
-        
-   ### 조현진
-     - 구축된 라즈베리 환경에 베이스라인 cv모델 구현 및 테스트
-     - gitgub에 사용 가능한 Dockerfile 올리기
-        
-   ### 반석, 나영준 
-     - 임베디드 시스템 최적화를 위한 onnx runtime 프레임워크로 베이스라인 cv모델 변환
-
-   ### 공통사항
-     - GIT을 적극적으로 사용하여 이슈관리
