--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-04-15.md
@@ -1,43 +1,0 @@
-## 교수님 면담 ##
-- 다음주 발표준비
-  한문장으로 뭘 하고 싶은지 설명이 되는게 좋다
-- 문제 정의 피드백
-- 풀고자 하는 문제에 대한 설명을 하기 위해선 관련기술 설명이 탁월하다
-   ADAS(자동차에 사용하는 기능)예시나
-  프라이버시 관련 문제(찾아볼 필요가 있음) 가 좋은 예시이다
-- 서버에서 돌아가는 알고리즘과 엣지에서 돌아가는 알고리즘의 차이에 대한 설명이 필요하다
-## 이번주 할일
-- 멘토링 자료준비 및 발표준비
-
-## 조현진
-- 멘토링 자료준비
-## 허남수
-- 자료준비
-## 서윤오
-- ppt준비
-## 반석
-- 발표준비
-## 나영준
-- raspi 와 docker 공부
-# 공통사항
-- github repo 관리를 위한 git 공부
-- 디버깅
-
-## 멘토링 ##
-- 진행 상황에 대한 피드백
-- 개발 상황들을 깃허브에 잘 남겼으면 좋겠다
-- 이슈를 하나보다는 나눠서 관리했으면 좋겠다
-- 이미 있는 이미지를 도커에 올리는 작업은 따로 관리를 했으면 좋겠다
-- Repository에 Wiki를 하나 만들어서 camera 이슈에 대한 것을 작성해라
-- Open CV에 연결하지 않고 카메라와 바로 연결해서 실행하는 예제를 테스트 해보자
-- 문서화가 중요하다
-- 모델 변환에 관한 답변
-- quantization을 할 때는 같은 모델을 할 때도 다른 경우가 있다.(모델 변환 시 노력에 따른 퍼포먼스가 나오는지에 관한 답변)
-- 여러 모델을 지월하는게 가장 돟다.
-- 처음 지원할 모델이 onnxruntime이라면 조금 더 범용성이 좋을것이라 생각했다.
-
-## 시험 기간 후 할 일 
-
-## 허남수 서윤오
-- Rasberry Pi 5에 Docker환경 구축부터 YOLO5 샘플코드 실행까지 했던 과정을 이슈로 정리
-
