--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-04-29.md
@@ -1,23 +1,0 @@
-## 교수님 면담
-- 개인 목표를 만들고, 이를 2주 후에 개별적으로 임의평가
-
-## 이번주 할일
-- openCV 데이터로 비전모델 결과 및 학습 데이터 도출
-- 도커 컨테이너 환경에서 picamera 모듈 임포트 및 프레임 확인
-- Docker-compose 실습 및 공부
-- 이슈관리
-
-## 조현진
-- openCV 데이터로 비전모델 결과 및 학습 데이터 도출
-
-## 허남수 & 서윤오
-- Docker-compose 실습 및 공부
-- Docker 안에서 yolo 환경 설정 및 테스트 했던 이슈 남기기
-
-## 반석 & 나영준
-- 도커 컨테이너 환경에서 picamera 모듈 임포트 및 프레임 확인
-
-# 공통사항
-- github repo 관리를 위한 git 공부
-- 디버깅
-
