--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-05-13.md
@@ -1,18 +1,0 @@
-## 이번주 할일
-- 현재 Github 자료 정리 및 문서화
-- TFlite와 ONNX로 작성된 pre-trained 모델 사용을 위한 환경구축
-- 커스텀 모델 학습환경 구축
-
-## 조현진 & 반석
-- TFlite와 ONNX로 작성된 pre-trained 모델 사용을 위한 환경구축
-
-## 허남수 & 서윤오
-- 커스텀 모델 학습환경 구축
-- 현재 Github 자료 정리 및 문서화
-
-## 나영준
-- CCA (Car Connect Accident) 
-
-# 공통사항
-- github repo 관리를 위한 git 공부
-- 디버깅
