--- test_fastAPI/실증적 sw프로젝트/주간회의록/2024-05-27.md
@@ -1,16 +1,0 @@
-## 이번주 할일
-
-## 허남수 & 서윤오
-- 현재 web 컨테이너에서 Flask 로컬 서버 구축
-- 컨테이너 내에서 webRTC를 이용하여 p2p 통신
-
-## 나영준 & 반석 & 조현진
-- Raspi5 - Docker 카메라 자원 할당 이슈 디버깅
-
-
-# 공통사항
-- github repo 관리를 위한 git 공부
-- 디버깅
-
-
-
